pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "django>=5.2.8",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
